sisu-wrapper/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sisu-wrapper"
version = "0.4.1"
authors = [
  { name="Ki Chun Tong", email="kichun.tong@aalto.fi" },
]
description = "Python library and REST API for Aalto University Sisu course data"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["sisu", "aalto", "courses", "api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.32.5",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "python-dotenv>=1.0.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.urls]
Homepage = "https://github.com/kctong529/sisukas"
Repository = "https://github.com/kctong529/sisukas/tree/main/sisu-wrapper"
Issues = "https://github.com/kctong529/sisukas/issues"

[tool.bumpver]
current_version = "0.4.1"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "bump version {old_version} -> {new_version}"
tag_message = "{new_version}"
